refactor: 카드 로컬 브랜드 관련 상수의 파일 경로 변경
